% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValues.R
\name{collectStatistics}
\alias{collectStatistics}
\title{Collect statistics}
\usage{
collectStatistics(
  ntimes,
  npop,
  monthL,
  monthP,
  monthE,
  monthR,
  monthM,
  monthO,
  smt,
  count,
  gemmonth,
  gemwinter,
  geewinter,
  geemonth
)
}
\description{
Collect statistics
}
