% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionModel.R
\name{correctionModel}
\alias{correctionModel}
\title{Correction Model Function}
\usage{
correctionModel(
  data,
  model = c("ranger", "bart", "xgboost", "nn", "svm", "lasso", "lm"),
  seed,
  ...
)
}
\arguments{
\item{data}{A data.frame containing the dataset to be used.}

\item{model}{Character. The model to be used for correction.
The available options are 'ranger', 'bart', 'xgboost', 'nn', 'svm', 'lasso', and 'lm'.
The default model is 'ranger'.}

\item{seed}{Numeric. The seed for reproducibility. Not used in this function but might be used in the 'getPredicts' function.}

\item{...}{Additional arguments to be passed to the model's function.}
}
\value{
This function returns the result of calling 'getPredicts' on the input data.
}
\description{
This function modifies the class of the input data.frame to match the
model specified and then attempts to generate predictions.
}
\examples{
\dontrun{
corrected_data <- correctionModel(data = mtcars, model = "xgboost", seed = 123)
}

}
