% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadData.R
\name{dataFrameData}
\alias{dataFrameData}
\title{Convert to data frame}
\usage{
dataFrameData(path)
}
\arguments{
\item{path}{File location.}
}
\value{
A data frame object.
}
\description{
Convert a excel file (or other extension) in a data frame.
}
