% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffOP.R
\name{diffOP}
\alias{diffOP}
\title{piner}
\usage{
diffOP(observed, predicted, quantile = 0.5)
}
\arguments{
\item{observed}{The observed value at the quantile}

\item{predicted}{The predicted value at the quantile}

\item{quantile}{The quantile where the difference should be calculated}
}
\value{
It returns a numeric value.
}
\description{
The diffOP function calculates the difference between two points of a cumulative curve at a given quantile.
}
\details{
Difference between the observed and the predicted at a quantile
}
