% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadData.R
\name{downloadData}
\alias{downloadData}
\title{Dowload data}
\usage{
downloadData(name, link, path, type = "xlsx", gdrive = TRUE)
}
\arguments{
\item{name}{A name for your file}

\item{link}{The data link.}

\item{path}{Location on the computer where the data will be saved.}

\item{type}{Type of file. Default is .xlsx.}
}
\value{
Returns a file saved in the designated folder.
}
\description{
Download data from google drive or other site.
}
