% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValues.R
\name{getFirstGeneration}
\alias{getFirstGeneration}
\title{First Geeneration}
\usage{
getFirstGeneration(npop, params, count, stumpTemp)
}
\arguments{
\item{npop}{population size}

\item{params}{values of parameters}

\item{stumpTemp}{stump temperature information}
}
\value{
First Generation information
}
\description{
Get first generation information.
}
