% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValues.R
\name{getStumpStats}
\alias{getStumpStats}
\title{Stump stats}
\usage{
getStumpStats(
  nd,
  temp,
  maxtemp,
  day,
  year,
  firstdisp,
  lastdisp,
  firstemrg,
  firstow,
  noow,
  cold,
  params,
  counter1
)
}
\arguments{
\item{nd}{data frame}

\item{temp}{values of the parameters}

\item{maxtemp}{jcnrjcr}

\item{day}{jcrnjr}

\item{year}{jjrv}

\item{firstdisp}{jnvjv}

\item{lastdisp}{cnencrej}

\item{firstemrg}{ncvenke}

\item{firstow}{vnevne}

\item{noow}{ncrenvje}

\item{cold}{vnrvnje}

\item{params}{vjnvjr}

\item{counter1}{jnvjrjnv}
}
\value{
Stump temperature information
}
\description{
jvsjvjsj
}
