% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStumpTemp.R
\name{getStumpTemp}
\alias{getStumpTemp}
\title{getStumpTemp}
\usage{
getStumpTemp(depth, nd, temp, th6)
}
\arguments{
\item{depth}{A single numeric value indicating the depth (in cm) at which to calculate stump temperatures.
Only 1 cm and 3 cm are accepted.}

\item{nd}{A single numeric value representing the number of days for which the calculation will be performed.}

\item{temp}{A numeric vector of length 'nd' containing daily temperature values in degrees Celsius.}

\item{th6}{A single numeric value representing the threshold temperature (in degrees Celsius)
below which the stump is considered cold.}
}
\value{
A list with the following components:
\itemize{
\item \code{s10}: A numeric vector of length 'nd' representing daily 10-day moving averages of temperature at 1 cm depth.
\item \code{s30}: A numeric vector of length 'nd' representing daily 30-day moving averages of temperature at 3 cm depth.
\item \code{cold}: A numeric vector of length 'nd' containing values of -1, 0, or 1, indicating whether the stump is cold
or not based on the 10-day moving average and threshold temperature.
\item \code{cumcold}: A numeric vector of length 'nd' containing cumulative cold values.
\item \code{stumpT}: A numeric vector of length 'nd' containing calculated stump temperatures based on the specified depth.
}
}
\description{
Calculate Stump Temperatures
}
\details{
This function calculates stump temperatures at specified depths and returns a list of attributes.
}
\examples{
depth <- 1
nd <- 10
temp <- c(5, 6, 7, 8, 9, 10, 11, 12, 13, 14)
th6 <- 6
getStumpTemp(depth, nd, temp, th6)
}
