% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module1.R
\name{module1}
\alias{module1}
\title{module1}
\usage{
module1(
  npop,
  data,
  params,
  stump_temp,
  first_generation,
  k,
  year,
  i,
  j,
  t0,
  gstyear,
  stagegg,
  rweevil,
  laval,
  stagel,
  t2
)
}
\arguments{
\item{npop}{A single numeric value representing the number of individuals in the population.}

\item{data}{A data frame containing daily climate data.}

\item{params}{A list containing various model parameters, such as threshold temperatures and time limits.}

\item{stump_temp}{A list containing stump temperature data.}

\item{first_generation}{A list containing information about the first generation of insects.}

\item{k}{A single numeric value representing the index for storing data in output vectors.}

\item{year}{A numeric vector of length 'nd' containing the year for each day.}

\item{i}{A single numeric value representing the current iteration of the outer loop.}

\item{j}{A single numeric value representing the current iteration of the inner loop.}

\item{t0}{A numeric vector of length 'npop' containing initial times for each individual in the population.}

\item{gstyear}{A numeric vector of length 'npop' containing the year of the first generation for each individual.}

\item{stagegg}{A numeric vector of length 'npop' containing the egg development time for each individual.}

\item{rweevil}{A numeric vector of length 'npop' containing the reproduction time for each individual.}

\item{laval}{A numeric vector of length 'npop' containing the larval development time for each individual.}

\item{stagel}{A numeric vector of length 'npop' containing the stage of development for each individual.}

\item{t2}{A numeric vector of length 'npop' containing the required degree days for larval development for
each individual.}
}
\value{
A list containing updated values for various insect development times, including:
\itemize{
\item Index values for storing data in output vectors (k and j).
\item Current day (today).
\item Larval development times (laval).
\item Start day for larval development (start).
\item Year of the first generation for each individual (gstyear).
\item Egg development time for each individual (stagegg).
\item Reproduction time for each individual (rweevil).
}
}
\description{
Module 1: Calculate Development Times for for Eggs and Larvae
This function calculates the time for eggs and larvae development of
the pine weevil.
}
