% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module2.R
\name{module2}
\alias{module2}
\title{module2}
\usage{
module2(prep, today, stump_temp, params, t3, t4, j)
}
\arguments{
\item{prep}{A numeric vector of length 'npop' containing the pre-pupal development times for each individual.}

\item{today}{A single numeric value representing the current day.}

\item{stump_temp}{A list containing stump temperature data.}

\item{params}{A list containing various model parameters, such as threshold temperatures and time limits.}

\item{t3}{A numeric vector of length 'npop' containing the required degree days for pre-pupal development for each individual.}

\item{t4}{A numeric vector of length 'npop' containing the required degree days for pupal development for each individual.}

\item{j}{A single numeric value representing the current iteration of the inner loop.}
}
\value{
A list containing updated values for various insect development times, including:
\itemize{
\item Current day (today).
\item Pre-pupal development times (prep).
\item Completion status for pupal development (cpupal).
}
}
\description{
Module 2: Calculate Pre-pupal and Pupal Development Times.
This function calculates the development times for the pre-pupal and pupal stages of
the pine weevil.
}
