% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module3.R
\name{module3}
\alias{module3}
\title{module3}
\usage{
module3(
  npop,
  params,
  stump_temp,
  today,
  prep,
  start,
  stagepp,
  stagepl,
  owr,
  emerge,
  i,
  j,
  gemtime,
  gemtime_new,
  gemwinter,
  gemmonth,
  year,
  gstyear,
  month,
  t0,
  rweevil
)
}
\arguments{
\item{npop}{A single numeric value representing the population size.}

\item{params}{A list containing various model parameters, such as degree days for emergence and overwintering.}

\item{stump_temp}{A list containing stump temperature and cold data.}

\item{today}{A single numeric value representing the current day.}

\item{prep}{A numeric vector of length 'npop' containing the pre-pupal development times for each individual.}

\item{start}{A single numeric value representing the starting day of the time series.}

\item{stagepp}{A numeric vector of length 'npop' containing the pre-pupal stage duration for each individual.}

\item{stagepl}{A numeric vector of length 'npop' containing the pupal stage duration for each individual.}

\item{owr}{A logical vector of length 'npop' containing the overwintering status for each individual.}

\item{emerge}{A numeric vector of length 'npop' containing the emergence times for each individual.}

\item{i}{A single numeric value representing the current iteration of the outer loop.}

\item{j}{A single numeric value representing the current iteration of the inner loop.}

\item{gemtime}{A numeric vector of length 'npop' containing the emergence times relative to the initial time (t0) for each individual.}

\item{gemtime_new}{A numeric vector of length 'npop' containing the new emergence times for each individual.}

\item{gemwinter}{A numeric vector of length 'npop' containing the number of winters passed before emergence for each individual.}

\item{gemmonth}{A numeric vector of length 'npop' containing the month of emergence for each individual.}

\item{year}{A numeric vector containing the year for each day.}

\item{gstyear}{A numeric vector of length 'npop' containing the generation start year for each individual.}

\item{month}{A numeric vector containing the month for each day.}

\item{t0}{A numeric vector of length 'npop' containing the initial time for each individual.}

\item{rweevil}{A numeric vector of length 'npop' containing the re-emergence status for overwintering individuals.}
}
\value{
A list containing updated values for various insect development times and overwintering status, including:
\itemize{
\item Current day (today).
\item Re-emergence status for overwintering individuals (rweevil).
\item Emergence times relative to the initial time (t0) for each individual (gemtime).
\item Number of winters passed before emergence for each individual (gemwinter).
\item Month of emergence for each individual (gemmonth).
\item Emergence times for each individual (emerge).
}
}
\description{
Module 3: Calculate Adult Emergence and Overwintering.
This function calculates the emergence times and overwintering status of
the pine weevil.
}
