% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module4.R
\name{module4}
\alias{module4}
\title{module4}
\usage{
module4(t7, today, params, stump_temp, month, j)
}
\arguments{
\item{t7}{A numeric vector of length 'npop' containing the required feeding times for each individual.}

\item{today}{A single numeric value representing the current day.}

\item{params}{A list containing various model parameters, such as degree days for emergence and overwintering.}

\item{stump_temp}{A list containing stump temperature and cold data.}

\item{month}{A numeric vector containing the month for each day.}

\item{j}{A single numeric value representing the current iteration of the inner loop.#'}
}
\value{
A list containing updated values for various insect maturation and hibernation times, including:
\itemize{
\item Current day (today).
\item Number of days spent feeding (feeddays).
}
}
\description{
Module 4: Calculate Maturation Feeding and Hibernation.
This function calculates the maturation feeding times and hibernation
status of the pine weevil.
}
