% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module5.R
\name{module5}
\alias{module5}
\title{module5}
\usage{
module5(params, today, maxtemp, stump_temp, t1, j)
}
\arguments{
\item{params}{A list containing various model parameters, such as degree days for emergence and overwintering.}

\item{today}{A single numeric value representing the current day.}

\item{maxtemp}{A numeric vector containing the maximum daily temperatures.}

\item{stump_temp}{A list containing stump temperature and cold data.}

\item{t1}{A numeric vector of length 'npop' containing the required egg development times for each individual.}

\item{j}{A single numeric value representing the current iteration of the inner loop.}
}
\value{
A list containing updated values for various insect dispersal, oviposition, and egg development times, including:
\itemize{
\item Current day (today).
\item Day of viable egg development (eggdev).
\item Index of selected viable egg (eggsel).
\item Vector of days for each egg (eggdays).
\item Vector of sorted indices for egg distribution (eggdist).
\item Vector of egg weights (eggwts).
}
}
\description{
Module 5: Calculate Dispersal, Oviposition, and Egg Development.
This function calculates the dispersal, oviposition, and egg development times
of the pine weevil.
}
