% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piner.R
\name{piner}
\alias{piner}
\title{PineR}
\usage{
piner(
  npop = 100,
  ntimes = 1,
  ngen = 5,
  data,
  species = "pine",
  depth = 2,
  owp = 100,
  params = TRUE,
  seed = 2021
)
}
\arguments{
\item{npop}{Population size.}

\item{ntimes}{Number of weather generator runs.}

\item{ngen}{Number of generations.}

\item{data}{Data containing the minimum, maximum and average temperatures, as well as day, month and year.}

\item{species}{Species of the three.}

\item{depth}{Stump depth (1 = 10cm, 2 = 20cm, 3 = 30cm)}

\item{owp}{Percentage that require overwintering}

\item{seed}{Seed for random number generator}
}
\value{
A list of objects.
}
\description{
A function to predict the life cycle of the Hylobius Abietis (Pine weevil).
}
