% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEmergence.R
\name{plotEmergence}
\alias{plotEmergence}
\title{plotEmergence
Plot Emergence Curve with Confidence Intervals.
This function creates a ggplot object visualizing the observed
and predicted emergence curves along with the confidence intervals
for the quantiles.}
\usage{
plotEmergence(quantiles, emergence, pal = c("steelblue", "firebrick"))
}
\arguments{
\item{quantiles}{A data frame containing quantile information
for the observed emergence curve.
 It should have columns 'day', 'qm' (median), 'ql' (lower quantile), and 'qu' (upper quantile).}

\item{emergence}{A data frame containing emergence probability
information for the predicted emergence curve. It should have columns 'day' and 'prob'.}

\item{pal}{A vector of length two representing the colours used to display the
observed (default: 'firebrick') and predicted (default: 'steelblue') values.}
}
\value{
A ggplot object representing the emergence curve plot.
}
\description{
plotEmergence
Plot Emergence Curve with Confidence Intervals.
This function creates a ggplot object visualizing the observed
and predicted emergence curves along with the confidence intervals
for the quantiles.
}
