% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMonths.R
\name{plotMonths}
\alias{plotMonths}
\title{plotMonths}
\usage{
plotMonths(model, pal = rainbow(12))
}
\arguments{
\item{model}{Output from the \code{piner} function.}

\item{pal}{The colour palette used to display the months.}
}
\value{
A ggplot object representing the stacked area plot of monthly probabilities
for different stages.
}
\description{
Create a Stacked Area Plot of Monthly Probabilities for Different Stages.
This function takes a simulation table and a data frame name as input, and
returns a ggplot object representing a stacked area plot of monthly probabilities
for different stages (L, P, E, M, and O) with a viridis color palette.
}
