% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantileValues.R
\name{quantileValues}
\alias{quantileValues}
\title{quantile_values}
\usage{
quantileValues(
  data,
  N,
  npop,
  ntimes,
  ngen = 1,
  species,
  depth,
  daysDiff,
  correction = TRUE
)
}
\arguments{
\item{data}{A data frame containing the input data.}

\item{N}{An integer representing the number of replicates to use in the simulation.}

\item{npop}{An integer representing the population size.}

\item{ntimes}{An integer representing the number of time steps in the simulation.}

\item{ngen}{An integer representing the number of generations in the simulation (default is 1).}

\item{species}{A character string specifying the species.}

\item{depth}{A numeric value representing the depth.}

\item{daysDiff}{An integer representing the difference in days between the start of the data and the start of the simulation.}

\item{correction}{A logical value indicating whether to apply a correction (default is TRUE).}
}
\value{
A data frame containing the quantile values for each day with columns: day, ql (lower quantile), qu (upper quantile), and qm (median quantile).
}
\description{
Calculate Quantile Values with Optional Correction. This function calculates the quantile values
for a given dataset and set of parameters, with the option to apply a correction.
It returns a data frame containing the quantile values for each day.
}
