% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadData.R
\name{readEmergence}
\alias{readEmergence}
\title{Read Emergence}
\usage{
readEmergence(data, colIndex = c(4, 6), percentual = TRUE)
}
\arguments{
\item{data}{Data frame of emergence}

\item{colIndex}{Columns of the days and the cumulative percentage}

\item{percentual}{If it is percentual or not}
}
\description{
Create a data frame with two columns: day of emergence and cumulative percentage
}
