% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modules.R
\name{runModule1}
\alias{runModule1}
\title{Module 1}
\usage{
runModule1(
  npop,
  data,
  params,
  count,
  stump_temp,
  first_generation,
  k,
  counter3
)
}
\arguments{
\item{npop}{kmk}

\item{data}{data}

\item{params}{params}

\item{count}{count}

\item{stump_temp}{stump_temp}

\item{first_generation}{first_generation}

\item{k}{k}

\item{counter3}{counter3}
}
\description{
Larval Development
}
