% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modules.R
\name{runModule2}
\alias{runModule2}
\title{Pupal}
\usage{
runModule2(params, prep, today, stump_temp, counter3)
}
\arguments{
\item{params}{kkfke}

\item{prep}{prep}

\item{today}{kvrkve}

\item{stump_temp}{vkvek}

\item{counter3}{mvremek}
}
\description{
Pupal Development
}
