% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modules.R
\name{runModule3}
\alias{runModule3}
\title{Emergence}
\usage{
runModule3(
  npop,
  data,
  today,
  count,
  params,
  stump_temp,
  owr,
  counter1,
  counter3
)
}
\arguments{
\item{npop}{kvvkekn}

\item{data}{data}

\item{today}{today}

\item{count}{count}

\item{params}{params}

\item{stump_temp}{stump_temp}

\item{owr}{owr}

\item{counter1}{counter1}

\item{counter3}{counter3}
}
\description{
Emergence
}
