% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modules.R
\name{runModule4}
\alias{runModule4}
\title{Maturation
 @description  Maturation}
\usage{
runModule4(data, count, today, params, stumpTemp, counter3)
}
\arguments{
\item{data}{data}

\item{count}{count}

\item{today}{today}

\item{params}{params}

\item{stumpTemp}{stumpTemp}

\item{counter3}{counter3}
}
\description{
Maturation
 @description  Maturation
}
