% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modules.R
\name{runModule5}
\alias{runModule5}
\title{Oviposition}
\usage{
runModule5(data, today, count, params, stump_temp, counter3)
}
\arguments{
\item{data}{data}

\item{today}{today}

\item{count}{count}

\item{params}{params}

\item{stump_temp}{stump_temp}

\item{counter3}{counter3}
}
\description{
Oviposition
}
