% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValues.R
\name{setCounters}
\alias{setCounters}
\title{Setting up the counters}
\usage{
setCounters(npop, ntimes)
}
\arguments{
\item{npop}{population size}

\item{ntimes}{number of times, simulations}
}
\value{
Empty vectors
}
\description{
Setting up the counters
}
