% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setParameters.R
\name{setParameters}
\alias{setParameters}
\title{setParameters}
\usage{
setParameters(species = NULL)
}
\arguments{
\item{species}{A character string representing the species of interest.
Currently supported species are "spruce" and "pine".}
}
\value{
A list containing various model parameters for different life stages,
 such as mean and standard deviation values for the duration of each stage,
 temperature thresholds, and feeding values. See associated vignette for a detailed
 list of all the parameters.
}
\description{
Set Parameters for Insect Life Cycle Model.
This function sets various parameters for running in the \code{piner} function.
}
\examples{
# Set parameters for the spruce species
spruce_params <- setParameters(species = "spruce")

# Set parameters for the pine species
pine_params <- setParameters(species = "pine")

}
