% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weatherData.R
\docType{data}
\name{weatherData}
\alias{weatherData}
\title{weatherData}
\format{
A data frame with 1096 rows and 6 columns:
\describe{
  \item{day}{Day of the year.}
  \item{month}{month of the year.}
  \item{year}{Year number.}
  \item{min_temp}{Minimum temperatures.}
  \item{max_temp}{Maximum temperatures.}
  \item{temp}{Temperature.}
  ...
}
}
\source{
Source of the dataset (e.g., a reference, URL, or citation)
}
\usage{
weatherData
}
\description{
Weather data for use in simulating pine weevil emergence.
}
\keyword{datasets}
