% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadData.R
\name{weightedTemp}
\alias{weightedTemp}
\title{Weight the data}
\usage{
weightedTemp(data, distances)
}
\arguments{
\item{data}{list of data frames}

\item{distances}{vector of distances associated with each element of the data frame list.}
}
\value{
A data frame with the date, minimum, maximum and average temperatures.
}
\description{
Inputting weighted averages
}
