% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValues.R
\name{collectStatistics}
\alias{collectStatistics}
\title{Collect statistics}
\usage{
collectStatistics(
  ntimes,
  npop,
  monthL,
  monthP,
  monthE,
  monthR,
  monthM,
  monthO,
  smt,
  count,
  gemmonth,
  gemwinter,
  geewinter,
  geemonth
)
}
\arguments{
\item{ntimes}{ntimes}

\item{npop}{npop}

\item{monthL}{monthL}

\item{monthP}{monthP}

\item{monthE}{monthE}

\item{monthR}{monthR}

\item{monthM}{monthM}

\item{monthO}{monthO}

\item{smt}{smt}

\item{count}{count}

\item{gemmonth}{gemmonth}

\item{gemwinter}{gemwinter}

\item{geewinter}{geewinter}

\item{geemonth}{geemonth}
}
\description{
Collect statistics
}
