% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectStatistics.R
\name{collectStatistics}
\alias{collectStatistics}
\title{collectStatistics}
\usage{
collectStatistics(
  ntimes,
  npop,
  monthL,
  monthP,
  monthE,
  monthR,
  monthM,
  monthO,
  SiteMeanTemp,
  count,
  gemmonth,
  gemwinter,
  geewinter,
  geemonth,
  firstdisp,
  lastdisp,
  firstemrg,
  firstow,
  noow
)
}
\arguments{
\item{ntimes}{A single numeric value representing the number of simulation runs.}

\item{npop}{A single numeric value representing the population size.}

\item{monthL}{A numeric value representing the sum of all individuals in the
larval stage across all months and simulation runs.}

\item{monthP}{A numeric value representing the sum of all individuals in the
pupal stage across all months and simulation runs.}

\item{monthE}{A numeric value representing the sum of all individuals in the
adult (emerging) stage across all months and simulation runs.}

\item{monthR}{A numeric value representing the sum of all individuals in the
reproducing stage across all months and simulation runs.}

\item{monthM}{A numeric value representing the sum of all individuals in the
migrating stage across all months and simulation runs.}

\item{monthO}{A numeric value representing the sum of all individuals in the
overwintering stage across all months and simulation runs.}

\item{SiteMeanTemp}{A single numeric value representing the sum of the mean
temperatures across all simulation runs.}

\item{count}{A list containing various counts of life cycle events from the simulation runs.}

\item{gemmonth}{A numeric vector containing the month of initial emergence for each individual.}

\item{gemwinter}{A numeric vector containing the number of winters before initial
emergence for each individual.}

\item{geewinter}{A numeric vector containing the number of winters before initial
oviposition for each individual.}

\item{geemonth}{A numeric vector containing the month of initial oviposition for each individual.}

\item{firstdisp}{Mean first dispersal day.}

\item{lastdisp}{Mean last dispersal day.}

\item{firstemrg}{Mean first emergence day.}

\item{firstow}{Mean first overwintering day.}

\item{noow}{Number of overwintering years.}
}
\value{
A list containing the processed statistics and proportions of various life cycle events, including:
\itemize{
\item Average proportion of individuals in each stage by month.
\item Mean temperature.
\item Mean days for various life cycle events (first dispersal, last dispersal, first emergence, first overwintering).
\item Percentage of the year with no overwintering.
\item Data frame with average proportions of individuals in each stage by month.
\item Proportions of initial emergence, initial oviposition, and emergence-oviposition across years.
\item Proportions of individuals completing the life cycle in 0, 1, 2, 3, or more than 3 years.
}
}
\description{
Collect and Process Life Cycle Statistics.
This function collects and processes life cycle statistics from simulation outputs,
calculating the averages and proportions of various life cycle events.
}
