% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFirstGeneration.R
\name{getFirstGeneration}
\alias{getFirstGeneration}
\title{getFirstGeneration}
\usage{
getFirstGeneration(npop, t0, t1, params, stump_temp, lavalst)
}
\arguments{
\item{npop}{A single numeric value representing the population size.}

\item{t0}{A numeric vector of length 'npop' containing the initial starting day for each individual.}

\item{t1}{A numeric vector of length 'npop' containing the thermal requirement (in degree days) for each individual.}

\item{params}{A list containing various model parameters, such as threshold temperatures and time limits.}

\item{stump_temp}{A list containing stump temperature data, including 10-day moving averages (s10).}

\item{lavalst}{A numeric vector of length 'npop' initialized with zeros, to store the calculated day when each individual reaches the larval stage.}
}
\value{
A list containing:
\itemize{
\item \code{today}: A single numeric value representing the last calculated day for the final individual in the population.
\item \code{lavalst}: A numeric vector of length 'npop' containing the calculated day when each individual reaches the larval stage.
}
}
\description{
Calculate First Generation Timings.
This function calculates the timings for the first generation of a
 population based on initial conditions, thermal requirements, and temperature data.
}
