% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStumpStats.R
\name{getStumpStats}
\alias{getStumpStats}
\title{getStumpStats}
\usage{
getStumpStats(
  nd,
  temp,
  maxtemp,
  day,
  year,
  firstdisp,
  lastdisp,
  firstemrg,
  firstow,
  noow,
  cold,
  params,
  cumcold,
  i
)
}
\arguments{
\item{nd}{A single numeric value representing the number of days.}

\item{temp}{A numeric vector of length 'nd' containing daily temperature data.}

\item{maxtemp}{A numeric vector of length 'nd' containing daily maximum temperature data.}

\item{day}{A numeric vector of length 'nd' containing the day of the year for each day.}

\item{year}{A numeric vector of length 'nd' containing the year for each day .}

\item{firstdisp}{A single numeric value representing the mean first dispersal day, updated during each iteration.}

\item{lastdisp}{A single numeric value representing the mean last dispersal day, updated during each iteration.}

\item{firstemrg}{A single numeric value representing the mean first emergence day, updated during each iteration.}

\item{firstow}{A single numeric value representing the mean first overwintering day, updated during each iteration.}

\item{noow}{A single numeric value representing the number of years with no overwintering, updated during each iteration.}

\item{cold}{A numeric vector of length 'nd' containing cold-related information for each day.}

\item{params}{A list containing various model parameters, such as threshold temperatures and time limits.}

\item{cumcold}{A numeric vector of length 'nd' containing cumulative cold-related information for each day.}

\item{i}{A single numeric value representing the current iteration of the outer loop.}
}
\value{
A list containing updated values for various stump temperature statistics, including:
\itemize{
\item Site mean temperature.
\item Cumulative cold-related information.
\item Cold day count (ccd).
\item Cold week count (cwd).
\item Mean first and last dispersal days.
\item Mean first emergence day.
\item Year, day, and cumulative cold information for calculating overwintering days.
\item Temporary date variable for overwintering calculations.
\item Number of years with no overwintering.
\item Mean first overwintering day.
}
}
\description{
Calculate Stump Temperature Statistics.
This function calculates various statistics related to stump temperature,
such as mean temperatures, dispersal days, emergence days, and overwintering days.
}
