% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValues.R
\name{getStumpTemp}
\alias{getStumpTemp}
\title{getStumpTemp}
\usage{
getStumpTemp(depth, nd, temp, th6)
}
\arguments{
\item{depth}{depth}

\item{nd}{length of temperature}

\item{temp}{mean temperature}

\item{th6}{param module 6}
}
\description{
Get the stump temperatures
}
