% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValues.R
\name{setParameters}
\alias{setParameters}
\title{Setting parameters}
\usage{
setParameters(species)
}
\arguments{
\item{species}{Tree species: pine or spruce}
}
\value{
A list of hyperparameters
}
\description{
Function to setting up the population parameters
}
