% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadData.R
\name{weatherData}
\alias{weatherData}
\title{Organize weather data}
\usage{
weatherData(data)
}
\arguments{
\item{data}{Weather data data frame}
}
\description{
Prepares the data in the format to be used in the function.
}
